// globals/types/user.type.ts
export interface UserSession {
  id?: string;
  _id?: string;
  user?: string;
  userId?: string;
  keycloakUserId?: string;
  keycloakRoles?: string[];
  keycloakGroups?: string[];
  userAgent?: string;
  ip?: string;
  isRevoked: boolean;
  expiresAt?: string;
  createdAt?: string;
  updatedAt?: string;
}